# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory System API Documentation",
        "description": "API documentation for the Inventory System",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/items": {
            "get": {
                "summary": "Get all items",
                "operationId": "5561acef6e4aef27354515aca3069b72",
                "responses": {
                    "200": {
                        "description": "List of items"
                    }
                }
            },
            "post": {
                "summary": "Create a new item",
                "operationId": "a4fb16eb9970fcb06a5d70ace3026f4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "item_name": {
                                        "type": "string"
                                    },
                                    "item_cid": {
                                        "type": "integer"
                                    },
                                    "item_sn": {
                                        "type": "string"
                                    },
                                    "item_spec": {
                                        "type": "string"
                                    },
                                    "item_save": {
                                        "type": "integer"
                                    },
                                    "item_price": {
                                        "type": "number"
                                    },
                                    "suggested_retail_price": {
                                        "type": "integer"
                                    },
                                    "item_note": {
                                        "type": "string"
                                    },
                                    "item_open": {
                                        "type": "integer"
                                    },
                                    "item_sort": {
                                        "type": "integer"
                                    },
                                    "item_mstock": {
                                        "type": "integer"
                                    },
                                    "item_type": {
                                        "type": "string"
                                    },
                                    "item_years": {
                                        "type": "string"
                                    },
                                    "item_holdmonth": {
                                        "type": "integer"
                                    },
                                    "item_outvyear": {
                                        "type": "string"
                                    },
                                    "item_predict": {
                                        "type": "integer"
                                    },
                                    "item_insertdate": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "item_editdate": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "item_barcode": {
                                        "type": "string"
                                    },
                                    "item_inbox": {
                                        "type": "integer"
                                    },
                                    "ppt_id": {
                                        "type": "integer"
                                    },
                                    "item_vcode": {
                                        "type": "string"
                                    },
                                    "item_size": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/v1/locations": {
            "get": {
                "summary": "Get all locations",
                "operationId": "9bbe66ccd9a8de4d94275b1f6c4be2b9",
                "responses": {
                    "200": {
                        "description": "List of locations"
                    }
                }
            },
            "post": {
                "summary": "Create a new location",
                "operationId": "3af24cfa81f4c045f6490e97e6653698",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "location_code": {
                                        "type": "string"
                                    },
                                    "location_name": {
                                        "type": "string"
                                    },
                                    "building_code": {
                                        "type": "string"
                                    },
                                    "floor_number": {
                                        "type": "string"
                                    },
                                    "floor_area_code": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "storage_type_code": {
                                        "type": "string"
                                    },
                                    "sub_area_code": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "position_code": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "current_stock": {
                                        "type": "integer"
                                    },
                                    "qr_code_data": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/v1/locations/{id}/floor-distribution": {
            "get": {
                "summary": "Get floor distribution for a specific location",
                "operationId": "e21fb94f495826cf968fed486b665728",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Floor distribution data"
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "400": {
                        "description": "Not a shelf type location"
                    }
                }
            }
        },
        "/api/v1/locations/{id}/items": {
            "get": {
                "summary": "Get items in a specific location",
                "operationId": "43dc0e6ad6672ed719fa43bb4cded225",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Items in location"
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            }
        },
        "/api/v1/locations/batch": {
            "post": {
                "summary": "Create multiple locations",
                "operationId": "a1e0fecbd453d6755e7eed8ee7e904f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "locations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "location_code": {
                                                    "type": "string"
                                                },
                                                "location_name": {
                                                    "type": "string"
                                                },
                                                "building_code": {
                                                    "type": "string"
                                                },
                                                "floor_number": {
                                                    "type": "string"
                                                },
                                                "floor_area_code": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "storage_type_code": {
                                                    "type": "string"
                                                },
                                                "sub_area_code": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "position_code": {
                                                    "type": "string"
                                                },
                                                "capacity": {
                                                    "type": "integer"
                                                },
                                                "current_stock": {
                                                    "type": "integer"
                                                },
                                                "qr_code_data": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "is_active": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Locations created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/movement-logs": {
            "get": {
                "summary": "Get movement logs with pagination and filters",
                "operationId": "551bcce5c71c1c45730094a74eb8d092",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "movement_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_range",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movement logs retrieved successfully"
                    }
                }
            }
        },
        "/api/v1/movement-logs/{id}": {
            "get": {
                "summary": "Get specific movement log",
                "operationId": "037dd76c54f07668f55be966d940ad01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movement log retrieved successfully"
                    },
                    "404": {
                        "description": "Movement log not found"
                    }
                }
            }
        },
        "/api/v1/movement-logs/qr-code/{qrCodeId}": {
            "get": {
                "summary": "Get movement history for specific QR code",
                "operationId": "214dd5ca3a8bcebc2c89011452e21cca",
                "parameters": [
                    {
                        "name": "qrCodeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movement history retrieved successfully"
                    }
                }
            }
        },
        "/api/v1/posin": {
            "get": {
                "summary": "Get all posin records with pagination and search",
                "operationId": "2e077300cd25930221f648435f20ddbe",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status filter",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of posin records"
                    }
                }
            },
            "post": {
                "summary": "Create a new posin record",
                "operationId": "6f587a50a9513d3715f0e50eeb695f82",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "_users_id": {
                                        "type": "integer"
                                    },
                                    "posin_sn": {
                                        "type": "string"
                                    },
                                    "posin_user": {
                                        "type": "string"
                                    },
                                    "posin_dt": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "posin_note": {
                                        "type": "string"
                                    },
                                    "posin_items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "itemtype": {
                                                    "type": "integer"
                                                },
                                                "item_id": {
                                                    "type": "integer"
                                                },
                                                "item_name": {
                                                    "type": "string"
                                                },
                                                "item_sn": {
                                                    "type": "string"
                                                },
                                                "item_spec": {
                                                    "type": "string"
                                                },
                                                "item_batch": {
                                                    "type": "string"
                                                },
                                                "item_count": {
                                                    "type": "integer"
                                                },
                                                "item_price": {
                                                    "type": "number"
                                                },
                                                "item_expireday": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "item_validyear": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Posin record created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/v1/posin/{id}/generate-us-purchase-order": {
            "patch": {
                "summary": "Generate US purchase order for a posin record",
                "operationId": "63ce20c1d88a80cb4b98d464c8b6f07e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "US purchase order generated successfully"
                    },
                    "404": {
                        "description": "Posin record not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/posin/{id}/items": {
            "get": {
                "summary": "Get posin items for a specific posin record",
                "operationId": "98c7fb4fb5724c0c7ebc75d309e0f21c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of posin items"
                    },
                    "404": {
                        "description": "Posin record not found"
                    }
                }
            }
        },
        "/api/v1/posin-items/{id}": {
            "delete": {
                "summary": "Delete a posin item",
                "operationId": "909e4b89f17424529b98e238676bb779",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posin item deleted successfully"
                    },
                    "404": {
                        "description": "Posin item not found"
                    }
                }
            }
        },
        "/api/v1/check-qr-generated/{posinitem_id}": {
            "get": {
                "summary": "Check if QR codes have been generated for a posin item",
                "operationId": "064918b85d3e0e460eaa6401d4ba67ba",
                "parameters": [
                    {
                        "name": "posinitem_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR generation status"
                    },
                    "404": {
                        "description": "Posin item not found"
                    }
                }
            }
        },
        "/api/v1/generate-qr-labels": {
            "post": {
                "summary": "Generate QR code labels for a posin item",
                "operationId": "9e0650e8ed9cc8dd47eddd5fdf7f66e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "item": {
                                        "type": "object"
                                    },
                                    "count": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "QR labels generated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/v1/qr-codes": {
            "get": {
                "summary": "Get all QR codes with pagination and search",
                "operationId": "f158335d7f511a916a495f0223b183d7",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status filter",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "generated",
                                "printed",
                                "used"
                            ]
                        }
                    },
                    {
                        "name": "location_id",
                        "in": "query",
                        "description": "Location ID filter",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of QR codes"
                    }
                }
            }
        },
...
        "/api/v1/qr-codes/by-zip-file/{zipFileName}": {
            "get": {
                "summary": "Get QR codes by ZIP file name",
                "operationId": "a6457e24bf96296227aff5325885e59f",
                "parameters": [
                    {
                        "name": "zipFileName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR codes in the ZIP file"
                    }
                }
            }
        },
...
    },
...
}